refactor: ServerBar
