+++ apps/dashboard/app/components/cards/server-card/card.tsx
@@ -11,23 +11,11 @@ import {
   type LucideIcon,
 } from "lucide-react";
 import type { Server } from ".";
+import ServerBar from "~/components/bars/ServerBar";
 import StatusBadge from "~/components/dashboard/StatusBadge";
 import { ServerStatusType } from "proto-gen-ts/daemon_pb";
 import { Card, CardContent, CardFooter, CardHeader } from "../../ui/card";
 import { Avatar, AvatarFallback, AvatarImage } from "../../ui/avatar";
-
-interface BarProps {
-  value: number;
-  max: number;
-  className?: string;
-}
-
-interface ResourceBarProps {
-  title: string;
-  value: number;
-  uiValue?: string;
-  max: number;
-}
 
 interface ServerCardProps {
   server: Server;
@@ -43,50 +31,6 @@ interface ServerCardLinkProps {
 }
 
 const serverCardTransition = "transition-all duration-300 ease-in-out truncate";
-
-const Bar: React.FC<BarProps> = ({ value, max, className }) => {
-  const percentage = Math.min((value / max) * 100, 100);
-
-  const getBarColor = () => {
-    if (percentage > 90) return "bg-rose-500";
-    if (percentage > 75) return "bg-amber-500";
-    return "bg-emerald-500";
-  };
-
-  return (
-    <div
-      className={cn(
-        "h-1.5 w-full overflow-hidden rounded-full",
-        "bg-slate-200 dark:bg-slate-800",
-        className
-      )}
-    >
-      <div
-        className={cn("h-full rounded-full", getBarColor())}
-        style={{ width: `${percentage}%` }}
-      />
-    </div>
-  );
-};
-
-const ResourceBar: React.FC<ResourceBarProps> = ({
-  title,
-  uiValue,
-  value,
-  max,
-}) => {
-  return (
-    <div className="space-y-1">
-      <div className="flex items-center justify-between text-xs">
-        <span className="text-slate-600 dark:text-slate-400">{title}</span>
-        <span className="font-medium text-slate-700 dark:text-slate-300">
-          {uiValue}
-        </span>
-      </div>
-      <Bar value={value} max={max} />
-    </div>
-  );
-};
 
 const ServerCardButton: React.FC<ServerCardLinkProps> = ({
   icon,
@@ -198,13 +142,13 @@ const ServerCard: React.FC<ServerCardProps> = ({ server, className }) => {
             </p>
           </CardHeader>
           <CardContent className="space-y-3">
-            <ResourceBar
+            <ServerBar
               title="CPU"
               uiValue={`${server.cpuUsage.toFixed(1)}%`}
               value={server.cpuUsage}
               max={100}
             />
-            <ResourceBar
+            <ServerBar
               title="Memory"
               uiValue={`${formatMemory(
                 server.memoryUsage.used
